backend/requirements.txt
faiss_cpu==1.9.0
fastapi==0.115.4
langchain==0.3.7
langchain_community==0.3.5
langchain_openai==0.2.6
numpy
openai==1.54.3
opencv_python==4.10.0.84
pydantic==2.9.2
pydub==0.25.1
pymupdf==1.24.13
tqdm==4.66.5
uvicorn==0.32.0
pillow